Code Factorization

+++ 16Product/main.py
@@ -1,7 +1,25 @@
 import os
 from product import Product
 
-askusertext = "\nEnter\n'1' to Add New Product\n'2' to Delete Product\n'3' to Edit Product\n'4' to View All Products\n'5' to Search Product\n'6' to Exit\n: "
+askuseroption = """
+Enter
+'1' to Add New Product
+'2' to Delete Product
+'3' to Edit Product
+'4' to View All Products
+'5' to Search Product
+'6' to Exit: 
+"""
+
+useroptionedit = """
+'1' for ID
+'2' for Product Name
+'3' for Quantity
+'4' for Price
+or Press any other key to Cancel Update: 
+"""
+
+
 fileLocation = "./16Product/product.csv"
 _isLoop = True
 
@@ -30,10 +48,10 @@ def searchProduct():
     searchID = askID()
     for data in readFile():
         if searchID == data.getID():
-            searchedData = data.getAllData()
+            searchedData = data
             break
         else:
-            searchedData = []
+            searchedData = None
     return searchedData
     
 
@@ -43,39 +61,41 @@ def addData():
     i=0
     while(user_input > i):
         print("\n")
-        id = input(f"Enter Product {i+1} ID: ")
+        p = Product()
+        p.id = input(f"Enter Product {i+1} ID: ")
         getID = [ data for data in readFile() if id == data.getID()]
         # print(getID)
         if getID == []:
-            n = input(f"Enter Product {i+1} Name: ")
-            q = input(f"Enter Product {i+1} Quantity: ")
-            p = input(f"Enter Product {i+1} Price: ")
+            p.name = input(f"Enter Product {i+1} Name: ")
+            p.quantity = input(f"Enter Product {i+1} Quantity: ")
+            p.price = input(f"Enter Product {i+1} Price: ")
             i+=1
-            writeFile(f"{id},{n},{q},{p}\n")
+            writeFile(f"{p.id},{p.name},{p.quantity},{p.price}\n")
         else: print("Can't have same id for two products.")
 
 def deleteData():
-    delID = askID()
-    delData = [ data for data in readFile() if delID == data.getID()]
-    isSure = input(f"Are you sure you want to delete product {delData[0].getAllData()} (Y/N): ").lower()
-    if isSure == 'y':
-        finalData = [ data for data in readFile() if delID != data.getID()]
-        # print(finalData)
-        os.remove(fileLocation)
-        for data in finalData: 
-            writeFile(f"{data.id},{data.name},{data.quantity},{data.price}\n")
-        print(f"{delData[0].name} has been deleted.")
-    else: print(f"{delData[0].name} isn't deleted.")
+    delData = searchProduct()
+    if delData != None:
+        isSure = input(f"Are you sure you want to delete product {delData.name} (Y/N): ").lower()
+        if isSure == 'y':
+            finalData = [ data for data in readFile() if delData.id != data.getID()]
+            # print(finalData)
+            os.remove(fileLocation)
+            for data in finalData: 
+                writeFile(f"{data.id},{data.name},{data.quantity},{data.price}\n")
+            print(f"{delData.name} has been deleted.")
+        else: print(f"{delData.name} isn't deleted.")
+    else: print("Can't Find Product to Delete.")
 
 def editProduct(value,type,editedProduct):
-    finalData = [ data for data in readFile() if editedProduct[0].id != data.getID() ]
-    editedProduct[0].updateData(value,type)
-    finalData.append(editedProduct[0])
+    finalData = [ data for data in readFile() if editedProduct.id != data.getID() ]
+    editedProduct.updateData(value,type)
+    finalData.append(editedProduct)
     # print(finalData)
     os.remove(fileLocation)
     for data in finalData: 
         writeFile(f"{data.id},{data.name},{data.quantity},{data.price}\n")
-    print(f"{editedProduct[0].id} is updated.")
+    print(f"{editedProduct.id} is updated.")
 
 def showProducts():
     print("ID\tProduct\t\tQuantity\tPrice\tTotal")
@@ -85,7 +105,7 @@ def showProducts():
 
 # try:
 while _isLoop:
-    value = int(input(askusertext))
+    value = int(input(askuseroption))
     if value == 1:
         addData()
 
@@ -93,21 +113,22 @@ while _isLoop:
         deleteData()
 
     elif value == 3:
-        editID = askID()
-        editData = [data for data in readFile() if editID == data.getID()]
-        isType = input(f"What do you want to edit in this product {editData[0].getAllData()}:\n'1' for ID\n'2' for Product Name\n'3' for Quantity\n'4' for Price\nor Press any other key to Cancel Update\n: ")
-        if isType == '1': value = input("Enter new ID: "); editProduct(value,isType,editData)
-        elif isType == '2': value = input("Enter new Product Name: "); editProduct(value,isType,editData)
-        elif isType == '3': value = input("Enter new Quantity: "); editProduct(value,isType,editData)
-        elif isType == '4': value = input("Enter new Price: "); editProduct(value,isType,editData)
-        else: print("Update is cancelled.")
+        editData = searchProduct()
+        if editData != None:
+            isType = input(f"What do you want to edit in this product {editData.getAllData()} {useroptionedit}")
+            if isType == '1': value = input("Enter new ID: "); editProduct(value,isType,editData)
+            elif isType == '2': value = input("Enter new Product Name: "); editProduct(value,isType,editData)
+            elif isType == '3': value = input("Enter new Quantity: "); editProduct(value,isType,editData)
+            elif isType == '4': value = input("Enter new Price: "); editProduct(value,isType,editData)
+            else: print("Update is cancelled.")
+        else: print("Can't Find Product to Edit.")
     
     elif value == 4:
         showProducts()
 
     elif value == 5:
         data = searchProduct()
-        if data != []:
+        if data != None:
             print(data[0],data[1],data[2],data[3])
         else:
             print("Can't Find the product")
